# Conflicts:
#	app/views/report/_checked.haml
#	config/locales/diaspora/en.yml
#	db/migrate/20221030224657_add_action_fields_to_reports.rb

+++ db/migrate/20221030224657_add_action_fields_to_reports.rb
@@ -2,7 +2,7 @@ class AddActionFieldsToReports < ActiveRecord::Migration[6.1]
 
   def change
     change_table :reports, bulk: true do |t|
-      t.numeric :reported_author_id, index: true
+      t.integer :reported_author_id, index: true
       t.string :action
     end
